Rename tests/ffi/module.rs to alias.rs to clarify purpose
